Refactor ProductsList into modules
